requirements.txt
pandas==1.2.3
openpyxl==3.0.7
python-docx==0.8.10
docx2pdf==0.1.7
openpyxl==3.0.7